[AbstractWorld] change name of milk_2 and cereal_2
